+++ back/luckybocky/src/main/java/com/project/luckybocky/feedback/dto/FeedbackReqDto.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class FeedbackReqDto {
 	private String feedbackContent;
-	private Integer feedbackRate;
+	private Byte feedbackRate;
 }
